Change asteroseismic sample to be in absolute magnitude space and verify classifications match.
Change the plot showing the asteroseismic sample to be in absolute magnitude
space. This time, spectroscopic log(g)s are unnecessary, so there is only the
asteroseismic sample overplotted on the hot sample.

Also made a plot showing that classification according to M_K is equivalent to
classification according to log(g). Splitting the asteroseismic sample up into
Hot, cool dwarfs, and cool subgiants results in the groups being in the same
locations in Teff/M_K space.