Separated calculations of Gaia radii for in-place or main-sequence.
We use two separate calculations to calculate Gaia radii. One uses the
intrinsic K-mag luminosity of the star. This would be appropriate for single
stars. However, for binaries, this will yield a radius which is unphysical. And
so as a result, to get a more reliable radius, we use the main-sequence
luminosity.